Improved faest_test.py overall

+++ faest_test.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
-# This script uses the prepare_nist.py script to export the
-# specified faest variants to the pqm4 project.
+# This script uses to test and benchmark specified faest variants
 
 import os
 import subprocess
@@ -9,6 +8,8 @@ import shutil
 import time
 
 from parse_massif import parse_and_write
+
+COMPILER = 'gcc'
 
 BASE_VARIANTS = ['128', '192', '256', 'em128', 'em192', 'em256']
 FAST_VARIANTS = [f'{v}f' for v in BASE_VARIANTS]
@@ -19,14 +20,19 @@ FILEPATH = os.path.dirname(os.path.realpath(__file__))
 TEST_BUILD_PATH = f'{FILEPATH}/test/build'
 TEST_RESULT_PATH = f'{FILEPATH}/test/results'
 TEST_RESULT_MASSIF_PATH = f'{TEST_RESULT_PATH}/massif'
+TEST_RESULT_CALLGRIND_PATH = f'{TEST_RESULT_PATH}/callgrind'
 TEST_FILE_PATH = f'{FILEPATH}/test/files'
 
 # TEST_FILE_NAME = 'faest_test.c'
-TEST_KEYGEN_NAME = 'faest_test_keygen.c'
-TEST_SIGN_NAME = 'faest_test_sign.c'
-TEST_VERIFY_NAME = 'faest_test_verify.c'
-TESTS = (('keygen', TEST_KEYGEN_NAME),
-         ('sign', TEST_SIGN_NAME), ('verify', TEST_VERIFY_NAME))
+TEST_KEYGEN_FILE_NAME = 'faest_test_keygen.c'
+TEST_SIGN_FILE_NAME = 'faest_test_sign.c'
+TEST_VERIFY_FILE_NAME = 'faest_test_verify.c'
+
+TEST_LIST = (('keygen', TEST_KEYGEN_FILE_NAME),
+             ('sign', TEST_SIGN_FILE_NAME), ('verify', TEST_VERIFY_FILE_NAME))
+TEST_NAMES = list(name for (name, _) in TEST_LIST)
+
+TOOL_NAMES = ['massif', 'callgrind']
 
 
 # Append 'faest_' and insert '_' between 'em' and '128f'
@@ -48,16 +54,18 @@ def parse_args():
                         help='Thread count. Set to 0 for max utilization (default: 1)')
     parser.add_argument('variants', nargs='*',
                         help='List of variants. If empty, all is run')
+    parser.add_argument('--tests', nargs='+',
+                        help='List of tests. If not specified, all is run. (Tests might depend on each other)')
     parser.add_argument('-s', '--slow', action='store_true', default=False,
                         help='Include slow variants if variants argument is empty (default: False)')
-    parser.add_argument('-v', '--verbose', action='store_true',
-                        default=False, help='Be verbose')
+    parser.add_argument('--tool', default=None,
+                        help='Select valgrind tool for profiling (massif or callgrind)')
     parser.add_argument('--no-openssl', action='store_true',
                         default=False, help='Disables OpenSSL optimization')
-    parser.add_argument('--no-massif', action='store_true',
-                        default=False, help='Disables massif memory profiler')
-    parser.add_argument('--no-rebuild', action='store_true',
+    parser.add_argument('--no-forced-rebuild', action='store_true',
                         default=False, help='Disables forced ninja rebuild of FAEST')
+    parser.add_argument('-v', '--verbose', action='store_true',
+                        default=False, help='Be verbose')
 
     args = parser.parse_args()
 
@@ -73,6 +81,24 @@ def parse_args():
         variants = args.variants
     variants = transform_variants(variants)
 
+    # Check tests
+    if args.tests:
+        for t in args.tests:
+            if t not in TEST_NAMES:
+                raise argparse.ArgumentTypeError(
+                    f'{t} is not a valid test. Only {TEST_NAMES} may be used')
+
+    if not args.tests:
+        tests = TEST_NAMES
+    else:
+        tests = args.tests
+
+    # Check tool
+    if args.tool and args.tool not in TOOL_NAMES:
+        raise argparse.ArgumentTypeError(
+            f'{t} is not a valid tool. Only {TOOL_NAMES} may be used')
+
+    # Threads
     threads = args.threads
     if threads <= 0:
         threads = os.cpu_count()
@@ -80,10 +106,11 @@ def parse_args():
 
     return {'threads': threads,
             'variants': variants,
+            'tests': tests,
+            'tool': args.tool,
             'verbose': args.verbose,
             'no-openssl': args.no_openssl,
-            'no-massif': args.no_massif,
-            'no-rebuild': args.no_rebuild}
+            'no-forced-rebuild': args.no_forced_rebuild}
 
 
 # Compiles the reference implementation
@@ -100,6 +127,13 @@ def compile_faest(verbose=False):
 
 def ensure_folder(path):
     os.makedirs(path, exist_ok=True)
+
+
+def ensure_result_folder(args):
+    if args['tool'] == 'massif':
+        ensure_folder(TEST_RESULT_MASSIF_PATH)
+    if args['tool'] == 'callgrind':
+        ensure_folder(TEST_RESULT_CALLGRIND_PATH)
 
 
 def copy(src, dst):
@@ -126,10 +160,9 @@ def compile(variant, program, name, verbose):
     copy(f'{TEST_FILE_PATH}/{program}',
          f'{TEST_BUILD_PATH}/{variant}/{program}')
 
-    compiler = 'gcc'
     link_path = f'{FILEPATH}/faest/build'
 
-    cmd = [compiler, f'-L{link_path}', '-lfaest', '-o',
+    cmd = [COMPILER, f'-L{link_path}', '-lfaest', '-o',
            name, 'crypto_sign.c', program]
     cwd = f'{TEST_BUILD_PATH}/{variant}'
     if verbose:
@@ -150,16 +183,20 @@ def simplify_name(name):
     return name
 
 
-def start_process(variant, name, no_massif):
-    cmd = []
+def tool_cmd(variant, name, args):
+    if args['tool'] == 'massif':
+        return ['valgrind', '--tool=massif', '--stacks=yes', '--threshold=0.01',
+                '--peak-inaccuracy=0.1', '--time-unit=B', '--detailed-freq=1', '--max-snapshots=1000',
+                f'--massif-out-file={TEST_RESULT_MASSIF_PATH}/{simplify_name(variant)}_{name}']
 
-    if not no_massif:
-        # Run with memory check
-        ensure_folder(TEST_RESULT_MASSIF_PATH)
-        cmd += ['valgrind', '--tool=callgrind', f'--callgrind-out-file={TEST_RESULT_MASSIF_PATH}/{simplify_name(variant)}_{name}']
-        #cmd += ['valgrind', '--tool=massif', '--stacks=yes', '--threshold=0.01', 
-        #        '--peak-inaccuracy=0.1', '--time-unit=B', '--detailed-freq=1', '--max-snapshots=1000',
-        #        f'--massif-out-file={TEST_RESULT_MASSIF_PATH}/{simplify_name(variant)}_{name}']
+    if args['tool'] == 'callgrind':
+        return ['valgrind', '--tool=callgrind', f'--callgrind-out-file={TEST_RESULT_CALLGRIND_PATH}/{simplify_name(variant)}_{name}']
+
+    return []
+
+
+def start_process(variant, name, args):
+    cmd = tool_cmd(variant, name, args)
 
     cmd += [f'{TEST_BUILD_PATH}/{variant}/{name}']
 
@@ -169,15 +206,6 @@ def start_process(variant, name, no_massif):
     # MacOS dynamic lib path
     env['DYLD_LIBRARY_PATH'] = f'{FILEPATH}/faest/build'
 
-    # measure time of process
-    #start = time.time()
-    #process = subprocess.Popen(cmd, cwd=f'{TEST_BUILD_PATH}/{variant}', stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, env=env, text=True)
-    #process.wait()
-    #end = time.time()
-    #print(f'{variant} {name} took {end - start} seconds\n')
-    #return process
-
-
     return subprocess.Popen(cmd, cwd=f'{TEST_BUILD_PATH}/{variant}', stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, env=env, text=True)
 
 
@@ -186,15 +214,12 @@ def run_all(args, name):
     threads = []
     done = []
 
-    # Hide cursor
-    print('\033[?25l', end='')
-
     # Keep specified amount of threads bussy
     while len(remaining) > 0 or len(threads) > 0:
         while len(remaining) > 0 and args['threads'] > len(threads):
             variant = remaining.pop()
             threads.append(
-                (variant, start_process(variant, name, args['no-massif'])))
+                (variant, start_process(variant, name, args)))
 
         # Print progress
         print('\r\033[96mDone:\033[0m {}, \033[96mRunning:\033[0m {}, \033[96mWaiting:\033[0m {}   '.format(
@@ -218,8 +243,8 @@ def run_all(args, name):
         [simplify_name(x[0]) for x in threads],
         [simplify_name(x) for x in remaining]), end='')
 
-    # Show cursor and newline
-    print('\033[?25h')
+    # Newline
+    print('')
 
     # Check return codes
     success = True
@@ -242,14 +267,33 @@ def compile_all(args, program, name):
         compile(v, program, name, args['verbose'])
 
 
+def set_openssl(enable):
+    if enable:
+        pattern = 'if false'
+        change_from = 'false'
+        change_to = 'openssl.found()'
+    else:
+        pattern = 'if openssl.found()'
+        change_from = 'openssl.found()'
+        change_to = 'false'
+
+    meson_config_file = f'{FILEPATH}/faest/meson.build'
+    with open(meson_config_file, 'r') as f:
+        lines = f.readlines()
+        for (i, l) in enumerate(lines):
+            if pattern in l:
+                lines[i] = l.replace(change_from, change_to)
+
+    with open(meson_config_file, 'w') as f:
+        f.writelines(lines)
+
+
 def disable_openssl():
-    os.rename(f'{FILEPATH}/faest/aes.c', f'{FILEPATH}/faest/aes_old.c')
-    shutil.copyfile(f'{TEST_FILE_PATH}/aes_no_openssl.c',
-                    f'{FILEPATH}/faest/aes.c')
+    set_openssl(False)
 
 
 def restore_openssl():
-    os.replace(f'{FILEPATH}/faest/aes_old.c', f'{FILEPATH}/faest/aes.c')
+    set_openssl(True)
 
 
 # ninja does not seem to always detect changes, use this to force rebuild (fx going from openssl to without)
@@ -258,21 +302,25 @@ def clean_faest():
                    stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, shell=True)
 
 
-def remove_results():
-    shutil.rmtree(TEST_RESULT_MASSIF_PATH, ignore_errors=True)
+def remove_results(args):
+    if args['tool'] == 'massif':
+        shutil.rmtree(TEST_RESULT_MASSIF_PATH, ignore_errors=True)
+    if args['tool'] == 'callgrind':
+        shutil.rmtree(TEST_RESULT_CALLGRIND_PATH, ignore_errors=True)
 
 
 if __name__ == '__main__':
     args = parse_args()
-    # TODO Remove Keccak optimizations (if any?)
-    # TODO Remove AVX optimizations (Why are they here? compiler complained about them..)
+
     if args['no-openssl']:
         print('Disabling OpenSSL')
         disable_openssl()
 
-    if not args['no-rebuild']:
+    if not args['no-forced-rebuild']:
         print('Forcing rebuild of FAEST')
-        clean_faest()  # TODO do this until ninja is behaves
+        clean_faest()  # NOTE do this while ninja does not behave
+
+    print('Compiling FAEST')
     compile_faest(args['verbose'])
 
     if args['no-openssl']:
@@ -280,10 +328,13 @@ if __name__ == '__main__':
         restore_openssl()
 
     print('Removing old results')
-    remove_results()
+    remove_results(args)
+    ensure_result_folder(args)
     copy_all(args)
 
-    for o, n in TESTS:
+    for o, n in TEST_LIST:
+        if o not in args['tests']:
+            continue
         print(f'[Testing {o}]')
 
         print('> Compiling variants')
@@ -292,8 +343,11 @@ if __name__ == '__main__':
         print(f'Running variants ({args["threads"]} thread(s))')
         run_all(args, o)
 
+    exit(0)
+
+    # TODO fix parsing or remove?
     print('Parsing results')
-    groups = [(o.title(), o) for o, _ in TESTS]
+    groups = [(o.title(), o) for o, _ in TEST_LIST]
 
     filename = f'results{"-no-openssl" if args["no-openssl"] else ""}.md'
     parse_and_write(TEST_RESULT_MASSIF_PATH,
